requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
Django @ file:///tmp/build/80754af9/django_1609784082548/work
djangorestframework==3.12.2
mysqlclient==2.0.3
psycopg2 @ file:///C:/ci/psycopg2_1608129576564/work
PyMySQL==1.0.2
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
wincertstore==0.2
